+++ CHANGELOG.md
@@ -24,6 +24,7 @@ For each minor and major version, there is a corresponding [milestone on GitHub]
 ### Fixed bugs
 
 - Retrieving bidirectional edges from the type graph didn't worked for both directions.
+- When checking the equality of custom types, the values for the same property might have different TypeScript types, since optional propeties might be set to `undefined`.
 
 
 ## v0.3.0 (2025-08-15)